% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_manipulation.R
\name{plotModelAccuracy}
\alias{plotModelAccuracy}
\alias{plotModelAccuracy.MRAmodel}
\alias{plotModelAccuracy.MRAmodelSet}
\title{Plot heatmaps of the model simulation against the data weighted by the error, as well as the log fold change for the data and the prediction}
\usage{
plotModelAccuracy(x, ...)

\method{plotModelAccuracy}{MRAmodel}(
  model_description,
  limit = Inf,
  show_values = TRUE,
  graphs = c("accuracy", "diff", "data", "simulation", "prediction"),
  selected_treatments = c(),
  selected_readouts = c(),
  name = "",
  regroup = "no"
)

\method{plotModelAccuracy}{MRAmodelSet}(
  model_description,
  limit = Inf,
  show_values = TRUE,
  graphs = c("accuracy", "data", "simulation"),
  selected_treatments = c(),
  selected_readouts = c(),
  name = "",
  side_by_side = TRUE
)
}
\arguments{
\item{model_description}{An MRAmodel object}

\item{limit}{An integer to force the limit of the heatmaps}

\item{show_values}{Whether the values should be printed in the heatmap boxes or not.}

\item{graphs}{Define which graphs should be plotted. "accuracy" for the residual as seen by the model, "qq" for the qqnorm plot of those residuals, "diff" for the delta log data-simulation, "data" for the log-fold change computed from the data, "simulation" for the log-fold change simulated by the model, "prediction" for the log-fold change that would be predicted without the blank correction.}

\item{selected_treatments}{A vector with the names of the subset of treatments that should be plotted}

\item{selected_readouts}{A vector with the names of the subset of readouts that should be plotted}

\item{name}{The name of the model, used as subtitle in the plot}

\item{regroup}{A value in c('inhib', 'stim') which indicates if the data should be regrouped by inhibition or stimulation in the heatmap}

\item{side_by_side}{For MRAmodelSet whether data and simulation should be plotted together for each submodel (TRUE) as opposed to all data together and all simulation together (FALSE)}
}
\value{
Invisibly, a list with elements 'mismatch', 'stim_data' and 'simulation' corresponding to the values plotted.
}
\description{
Plot heatmaps of the model simulation against the data weighted by the error, as well as the log fold change for the data and the prediction

Plot model accuracy for MRAmodel

Plot accuracy of all submodels of a modelset
}
\seealso{
createModel, importModel

Other Model plots: 
\code{\link{plotModelGraph}()},
\code{\link{plotModelScores.MRAmodel}()},
\code{\link{plotModelScores}()},
\code{\link{plotParameters}()},
\code{\link{plotResiduals}()}
}
\author{
Mathurin Dorel \email{dorel@horus.ens.fr}
}
\concept{Model plots}
