% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model_manipulation.R
\name{accuracyPlot}
\alias{accuracyPlot}
\title{Plots heatmaps of the model prediction against the data, weighted by the error, as well as the log fold change data and the prediction}
\usage{
accuracyPlot(model_description)
}
\arguments{
\item{model_description}{A list describing the model, as the one produced by createModel or importModel}
}
\value{
Nothing
}
\description{
Plots heatmaps of the model prediction against the data, weighted by the error, as well as the log fold change data and the prediction
}
\author{
Mathurin Dorel \email{dorel@horus.ens.fr}
}
\seealso{
plotModelPrediction, createModel
}

