% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_data_generation.R
\name{addLayerLinks}
\alias{addLayerLinks}
\title{Generate a list of links between two layers with a probability of link between two nodes of p_link, with possibly at least one link to each node of the bottom layer}
\usage{
addLayerLinks(links, lup, ldown, p_link = 0.2, connect_bottom = TRUE)
}
\arguments{
\item{links}{Initial list of links}

\item{lup}{List of nodes of the upper layer}

\item{ldown}{List of nodes of the lower layer}

\item{p_link}{Probability of linking two nodes between the layer}

\item{connect_bottom}{Whether each node of the lower layer must have at least on link}
}
\description{
Generate a list of links between two layers with a probability of link between two nodes of p_link, with possibly at least one link to each node of the bottom layer
}
\seealso{
Other toy data generation: 
\code{\link{createSimulation}()},
\code{\link{generateToyDesign}()},
\code{\link{generateToyNetwork}()}
}
\concept{toy data generation}
