% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_manipulation.R
\name{addLink}
\alias{addLink}
\title{add Link routine}
\usage{
addLink(new_link, adj, rank, init_residual, model, initial_response, expdes,
  data, model_structure, sample, verbose = F)
}
\arguments{
\item{new_link}{integer link whose addition is to be tested}

\item{adj}{integer matrix original adjacency matrix excluding the new_link}

\item{init_residual}{numeric sum-squared error of original network}

\item{model}{MRAmodel object of original network}

\item{initial_response}{list containing the local_response matrix and inhibitor strength of original network}

\item{expdes}{design object of MRAmodel object}

\item{data}{data object of MRAmodel object}

\item{model_structure}{structure object of MRAmodel object}

\item{sample}{numeric vector containing all starting values for new_link}
}
\description{
add Link routine
}

