% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model.R
\name{addVariableParameters}
\alias{addVariableParameters}
\title{Extend a MRAmodelSet with relevant variable parameters}
\usage{
addVariableParameters(original_modelset, nb_cores = 0,
  max_iterations = 0, nb_samples = 100, accuracy = 0.95,
  method = "geneticlhs", notVariable = c())
}
\arguments{
\item{original_modelset}{An MRAmodelSet object}

\item{nb_cores}{Number of cores to use for the refitting with the new variable parameters}

\item{max_iterations}{Maximum number of variable parameters to add (if 0 takes as many as possible)}

\item{nb_samples}{Number of samples to generate to fit the new variable parameters}

\item{accuracy}{Cutoff probability for the chi^2 test}

\item{method}{Name of the LHS method to use}

\item{notVariable}{Parameters that should not be varied, either a numeric or character vector identifying the parameters from 'modelset$model$getParametersNames()$names', defaults to 'c()'.}
}
\value{
An updated MRAmodelSet with the new parameter sets
}
\description{
See if a MRAmodelSet requires some parameters to be variable among models to explain the variations
}
\author{
Mathurin Dorel \email{dorel@horus.ens.fr}

Bertram Klinger \email{bertram.klinger@charite.de}
}
