% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_set.R
\name{compareModels}
\alias{compareModels}
\title{Compare networks}
\usage{
compareModels(files)
}
\arguments{
\item{files}{A list of '.mra' files}
}
\value{
matrix of parameters
}
\description{
Compare the parameters of one network for different conditions (cell line, perturbations, ...)
Missing links are indicated by NA.
}
\author{
Bertram Klinger \email{bertram.klinger@charite.de}
}
