% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_set.R
\name{compareParameters}
\alias{compareParameters}
\alias{compareParameters.MRAmodelSet}
\alias{compareParameters.modelGroup}
\title{Compare networks}
\usage{
compareParameters(multi_model, ...)

\method{compareParameters}{MRAmodelSet}(modelset, do_plot = TRUE)

\method{compareParameters}{modelGroup}(modelgroup, do_plot = TRUE)
}
\arguments{
\item{modelset}{A MRAmodelSet object}

\item{do_plot}{Whether the matrix of parameters should be plotted}
}
\value{
Invisibly a matrix of parameters with models as column names
}
\description{
Compile the parameters of a model set in one matrix and plot the results with highlighting of the variable links by rowwise scaling.
}
\author{
Bertram Klinger \email{bertram.klinger@charite.de}

Mathurin Dorel \email{dorel@horus.ens.fr}
}
