% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mra_model.R
\name{computeFitScore}
\alias{computeFitScore}
\title{Compute the fitting score of a model}
\usage{
computeFitScore(mra_model)
}
\description{
Compute the fitting score of the model (fraction of the variation in the data explained by the network)
Do the computation for each measured node and for the network
}

