% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model_set.R
\name{createDataSet}
\alias{createDataSet}
\title{Import multiple files}
\usage{
createDataSet(model_links, data_list, basal_file, cores = 1, inits = 1000,
  init_distribution = F, method = "default")
}
\arguments{
\item{data_list}{Name of the file containing the list of the files for the data without the extension}
}
\description{
Import multiple files
}

