% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model.R
\name{createModelSet}
\alias{createModelSet}
\title{Build an MRAmodelSet}
\usage{
createModelSet(model_links, basal_nodes, csv_files, var_files = c(),
  nb_cores = 1, inits = 1000, perform_plots = F, method = "geneticlhs",
  unused_perturbations = "")
}
\description{
Build and fit an MRAmodelSet, which consists of the simultaneous fitting of several MRA models
}
\author{
Mathurin Dorel \email{dorel@horus.ens.fr}
}

