% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_comparison.R
\name{getDirectPaths}
\alias{getDirectPaths}
\alias{getDirectPaths.MRAmodel}
\alias{getDirectPaths.MRAmodelSet}
\alias{getDirectPaths.modelGroup}
\title{Compute direct paths for a model}
\usage{
getDirectPaths(mra_model, non_stop_nodes = c())

\method{getDirectPaths}{MRAmodel}(mra_model, non_stop_nodes = c())

\method{getDirectPaths}{MRAmodelSet}(mra_model, non_stop_nodes = c())

\method{getDirectPaths}{modelGroup}(mra_model, non_stop_nodes = c())
}
\arguments{
\item{mra_model}{An MRAmodel or MRAmodelSet object}

\item{non_stop_nodes}{A list of nodes that should not begin or end a path, usefull to compare models with different basal activities. If the node is only present as a first or last node of a path, the corresponding paths will be deleted.}
}
\value{
(MRAmodel) A list of lists indexed by the path name. Each sublist has fields path, value, hv and lv which represent respectively the path equation, the value of the path, the upper bound of the path and the lower bound of the path.

(modelGroup and MRAmodelSet) A list of matrices with the direct path values and the limits of the 95% confidence interval
}
\description{
Compute the direct paths for a model and their confidence intervals by combining the paths containing links with negative exponents with paths containing the same link with a positive exponent.
}
