% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_export.R
\name{exportModel}
\alias{exportModel}
\title{Exports the model in a file}
\usage{
exportModel(model_description, file_name = "mra_model", export_data = FALSE)
}
\arguments{
\item{model_description}{An MRAmodel object}

\item{file_name}{Name of the output file}

\item{export_data}{Whether the CVs should be incorporated to the .mra file or not}
}
\value{
Nothing
}
\description{
Export an MRAmodel object in a .mra file
}
\seealso{
importModel, rebuildModel
}
\author{
Mathurin Dorel \email{dorel@horus.ens.fr}
}
