% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model.R
\name{extraction}
\alias{extractBasalActivity}
\alias{extractMIDAS}
\alias{extractStructure}
\alias{extraction}
\title{Extraction of object from file or object}
\usage{
extractStructure(to_detect, names = "", adj = FALSE)

extractBasalActivity(to_detect)

extractMIDAS(to_detect)
}
\arguments{
\item{to_detect}{An object to check or a string. A string is interpreted as a path to a file containing a matrix compatible with the target object.}

\item{adj}{Boolean, whether a 2 or 3 columns matrix in to_detect should be interpreted as an adjacency matrix}

\item{name}{Name for the structure. Extracted automatically from the file name if used.}
}
\value{
'extractStructure' returns a C++ object of class 'ModelStructure'

'extractBasalActivity' returns vector of character

'extractMIDAS' returns a matrix or a data.frame under MIDAS format
}
\description{
Extraction of object from file or object

'extractStructure' detects the format of the structure file and extract the structure of the network. It expects a matrix with 2 or 3 columns (adjacency list) or a square matrix (adjacency matrix)

'extractBasalActivity' detects if the a list of names or a file perform the extraction if necessary. It expects a vector or an object convertible to a vector.

'extractMIDAS' detects if the argument is a MIDAS dataset or the name of a file containing such dataset. It expects a matrix in MIDAS format (See Details).
}
\details{
A MIDAS format matrix is a matrix with specific column names:
\itemize{
 \item { One column 'ID:type' containing the type of experiment of each row:\itemize{
   \item {'c' or 'control' for control experiment with no treatment}
   \item {'blank' for blank experiment with no treatment}
   \item {'t' for experiment with a treatment}
 }}
 \item { At least one column 'TR:STIMULATION' or 'TR:INHIBITIONi' containing 0 or 1 whether the perturbation is applied for the experiment in each row }
 \item { One column 'DA:all' or the relevant number of columns 'DA:PERTURBATION' containing the time of the perturbation (Not used in STASNet) }
 \item { At least one column with column name 'DV:READOUT' containing the measured values for each readout }
}
}
\seealso{
\code{\link{readMIDAS}}
}

