% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_set.R
\name{extractSubmodels}
\alias{extractSubmodels}
\title{Extract individual models from an MRAmodelSet}
\usage{
extractSubmodels(modelset)
}
\arguments{
\item{modelset}{An MRAmodelSet object}
}
\value{
A list of MRAmodel objects
}
\description{
Extract individual models from an MRAmodelSet where they were fitted together and return them in a list where the indices are the names of each submodel.
}
\author{
Mathurin Dorel \email{dorel@horus.ens.fr}
}
