% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitmodel.R
\docType{package}
\name{fitmodel}
\alias{fitmodel}
\alias{fitmodel-package}
\title{Fit and simulate model from perturbation data}
\description{
Fits and simulates network model derived from perturbation data using Modular Response Analysis and Profile Likelihood
}
\details{
This package provides functions to fit, simulate and examine a model network from perturbation data. It uses the Modular Response Analysis method described in "Network quantification of EGFR signaling unveils potential for targeted combination therapy, Klinger et al." to build the model.

For a complete list of functions and classes, use 'library(help=fitmodel)'.

For a walk-through as to how the package can be used to fit a model, see the vignette.
}
\author{
Bertram Klinger \email{bertram.klinger@charite.de}, Mathurin Dorel \email{dorel@horus.ens.fr}, Nils Blüthgen \email{nils.bluethgen@charite.de}
}
\references{
Klinger et al. (2013) Network quantification of EGFR signaling unveils potential for targeted combination therapy. Molecular Systems Biology.
}

