% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_data_generation.R
\name{generateToyDesign}
\alias{generateToyDesign}
\title{Generate an experimental design for a toy network}
\usage{
generateToyDesign(network, nmes = 4, ninh = 2, stim_combo = 1, inhib_combo = 1)
}
\arguments{
\item{network}{A 2-columns matrix. The toy network as an adjacency list}

\item{nmes}{Number of measured nodes (chosen randomly)}

\item{ninh}{Number of inhibited nodes (chosen randomly)}

\item{stim_combo}{Number of simultaneous stimulations}

\item{inhib_combo}{Number of simultaneous inhibitions}
}
\value{
An experimental design object
}
\description{
Generate an experimental design for a toy network, with all stimulus applied, some nodes inhibited and some nodes measured chosen randomly.
}
\seealso{
\code{\link{getCombinationMatrix}}, \code{\link{generateToyNetwork}}

Other toy data generation: 
\code{\link{addLayerLinks}()},
\code{\link{createSimulation}()},
\code{\link{generateToyNetwork}()}
}
\concept{toy data generation}
