% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_model.R
\name{getCombinationMatrix}
\alias{getCombinationMatrix}
\title{Create a perturbation matrix}
\usage{
getCombinationMatrix(
  perturbations,
  inhib_combo = 2,
  stim_combo = 1,
  byStim = T
)
}
\arguments{
\item{perturbations}{A vector with the name of the perturbation, either NODE for a stimulation a node, or NODEi for its inhibition}

\item{inhib_combo}{Number of inhibitions to use simultaneously in each perturbation}

\item{stim_combo}{Number of stimulations to use simultaneously in each perturbation}

\item{byStim}{Whether the perturbations should be ordered according to the stimulations or the inhibitions}
}
\value{
A perturbation matrix with the names of the nodes as columnnames
}
\description{
Create the perturbation matrix for a set of perturbations, building all n-combinations of stimulators with all m-combinations of inhibitors. Add the cases with only stimulations and only inhibitions.
}
\seealso{
Other simulation: 
\code{\link{plotSimulation}()},
\code{\link{simulateModel}()}
}
\author{
Mathurin Dorel \email{dorel@horus.ens.fr}
}
\concept{simulation}
