% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_comparison.R
\name{getDirectPaths}
\alias{getDirectPaths}
\alias{printDirectPaths}
\title{Compute direct paths for a model}
\usage{
getDirectPaths(mra_model)

printDirectPaths(mra_model, digits = 2)
}
\arguments{
\item{mra_model}{An MRAmodel object}

\item{digits}{Number of digits to display for the confidence interval values}
}
\value{
A list of lists indexed by the path name. Each sublist has fields path, value, hv and lv which represent respectively the path equation, the value of the path, the upper bound of the path and the lower bound of the path.
}
\description{
Compute the direct paths for a model and their confidence intervals by combining the paths containing links with negative exponents with paths containing the same link with a positive exponent.

Print the direct paths with confidence interval
}

