% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_model.R
\name{getExperimentalDesign}
\alias{getExperimentalDesign}
\title{Generate an ExperimentalDesign object}
\usage{
getExperimentalDesign(
  model.structure,
  stim.nodes,
  inhib.nodes,
  measured.nodes,
  stimuli,
  inhibitor,
  basal.activity = c()
)
}
\arguments{
\item{model.structure}{A ModelStructure object}

\item{stim.nodes}{A character vector for the stimulated nodes}

\item{inhib.nodes}{A character vector for the inhibited nodes}

\item{measured.nodes}{A character vector listing the measured nodes}

\item{stimuli}{A matrix indicating which stimulation (column) is used for each treatment (row)}

\item{inhibitor}{A matrix indicating which inhibition (column) is used for each treatment (row)}

\item{basal.activity}{A character vector for the nodes with basal activity}
}
\value{
An ExperimentalDesign object
}
\description{
Generate an ExperimentalDesign object from the relevant input data after performing some sanity checks
}
