% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mra_model.R
\name{getLinksList}
\alias{getLinksList}
\title{Get the list of links}
\usage{
getLinksList(mra_model)
}
\arguments{
\item{mra_model}{An MRAmodel object.}
}
\value{
A data.frame with 2 columns, IN and OUT, corresponding to a directed adjacency list.
}
\description{
Get the structure of the model network as an adjacency matrix
}
