% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_manipulation.R
\name{getModelError}
\alias{getModelError}
\title{Compute the error of the model}
\usage{
getModelError(mra_model)
}
\arguments{
\item{mra_model}{An MRAmodel object}
}
\value{
A list with the simulation, the mismatch between the simulation and the data, and the residual of the fit
}
\description{
Compute the error of the model
}
