% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mra_model.R
\name{getParametersNames}
\alias{getParametersNames}
\title{Return the paths corresponding to each parameter}
\usage{
getParametersNames(mra_model)
}
\arguments{
\item{mra_model}{An MRAmodel object.}
}
\value{
A vector containing the names of the parameters
}
\description{
Return the paths corresponding to each parameter
}

