% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mra_model.R
\name{getParametersValues}
\alias{getParametersValues}
\title{Get the annotated parameters.}
\usage{
getParametersValues(mra_model)
}
\arguments{
\item{mra_model}{An MRAmodel object.}
}
\value{
A vector containing the values of the parameters, with the simplified path names as name.
}
\description{
Get the parameter values.
}
