% Generated by roxygen2 (4.0.1): do not edit by hand
\name{importModel}
\alias{importModel}
\title{Import model from a file}
\usage{
importModel(file_name)
}
\arguments{
\item{file_name}{Name of the .mra file}
}
\value{
An MRAmodel object
}
\description{
Import an MRAmodel object from a .mra file
}
\author{
Mathurin Dorel \email{dorel@horus.ens.fr}
}
\seealso{
exportModel, rebuildModel
}

