% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{log_norm_sd}
\alias{log_norm_sd}
\title{Log-normal standard deviation}
\usage{
log_norm_sd(xx, na.rm = TRUE)
}
\description{
Compute the standard deviation of a set following a log-normal distribution
}
