% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_export.R
\name{midasFromData}
\alias{midasFromData}
\title{Write a MIDAS file from a human readable dataset}
\usage{
midasFromData(data, fname)
}
\arguments{
\item{data}{Matrix where the rownames are the treatment separated by + signs and the column names are the readouts}
}
\description{
Write a MIDAS file from a human readable dataset
}
