% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_models.R
\name{modelGroup}
\alias{modelGroup}
\title{Constructor for modelGroup objects}
\usage{
modelGroup(models = list())
}
\arguments{
\item{models}{A list of models to use in the modelGroup object}
}
\value{
A modelGroup object
}
\description{
Construct a modelGroup object, which contains several mraModels and provides methods to compare them
}
