% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mra_model.R
\name{MRAmodel}
\alias{MRAmodel}
\title{Constructor for MRAmodel objects}
\usage{
MRAmodel(model = NULL, design = NULL, structure = NULL,
  basal = matrix(), data = matrix(), cv = matrix(),
  parameters = vector(), bestfit = NA, basefit = NA, name = "",
  infos = c(), param_range = list(), lower_values = c(),
  upper_values = c())
}
\arguments{
\item{model}{An object of class Model}

\item{design}{An object of class ExperimentalDesign}

\item{structure}{An object of class ModelStructure}

\item{basal}{A matrix describing the basal activity of the nodes in the network}

\item{data}{An object of class Data}

\item{cv}{A matrix containing the coefficient of variation of the data used to build the model}

\item{parameters}{A vector containing the values of the parameters of the model}

\item{bestfit}{The residual chi-2 value associated with the best fit}

\item{basefit}{The chi-2 value associated with the raw data (no fit)}

\item{name}{Name of the model}

\item{infos}{Extra information on the model}

\item{param_range}{Alternative parameters sets for the model}

\item{upper_values}{Upper bound on the values of the parameters}

\item{lower_value}{Lower bound on the values of the parameters}
}
\value{
An MRAmodel object
}
\description{
Construct an MRAmodel objects, which contains all the informations recquired to simulate the model, and the data used to build it.
The model is built using perturbation datas and an input network.
}
\seealso{
\code{\link{createModel}}
}

