% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_likelihood.R
\name{niplotPL}
\alias{niplotPL}
\title{Plots the functionnal relations between each non identifiable parameter and the profile likelihood of all parameters
and redirect it in a pdf file.}
\usage{
niplotPL(profiles, data_name = "default", folder = "./")
}
\arguments{
\item{profiles}{A list of the likelihood profiles of the parameters of the model}

\item{data_name}{Name for the output pdf file}

\item{folder}{Path to the folder for the output pdf file (must end with /)}
}
\value{
Nothing
}
\description{
Plots the functionnal relations between each non identifiable parameter and the profile likelihood of all parameters
and redirect it in a pdf file.
}
\seealso{
\code{\link{profileLikelihood}}
}
\author{
Mathurin Dorel \email{dorel@horus.ens.fr}
}
