% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.r
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot a customized heatmap with}
\usage{
plot_heatmap(mat, main = "", stripOut = 0.05, lim = Inf, fixedRange = F,
  col = colorRampPalette(c("deepskyblue", "white", "red1")),
  textCol = "gray10")
}
\arguments{
\item{mat}{a numeric matrix that should be plotted}

\item{main}{a character string representing the title to be passed}

\item{stripOut}{numeric between [0 1] to determine the range of colors (excluding outliers from distorting the colormap)}

\item{lim}{a single number indicating the maximal range [-lim lim] for the color map}

\item{fixedRange}{logical inidicating whether the range of lim should at all means be kept (ensures comparability with other heatmaps)}

\item{col}{colorRampPalette object indicating the colormap}

\item{textCol}{string denoting the color of inset text}
}
\value{
graphical object that can be plotted with print() function (optional)
}
\description{
Plot a customized heatmap with
}
\author{
Bertram Klinger \email{bertram.klinger@charite.de}
}

