% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatmap.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Plot a customized heatmap with a symmetrical scale that is not stretched by extreme values}
\usage{
plotHeatmap(mat, main = "", lim = Inf, fixedRange = F, stripOut = 0.05,
  col = colorRampPalette(c("deepskyblue", "white", "red1")),
  textCol = "gray10")
}
\arguments{
\item{mat}{a numeric matrix that should be plotted}

\item{main}{a character string representing the title to be passed}

\item{lim}{a single number indicating the maximal range [-lim lim] for the color map}

\item{fixedRange}{logical indicating whether the range of lim should at all means be kept (ensures comparability with other heatmaps)}

\item{stripOut}{numeric between [0 1] to determine the range of colors (excluding outliers from distorting the colormap)}

\item{col}{colorRampPalette object indicating the colormap}

\item{textCol}{string denoting the color of inset text}
}
\value{
Nothing
}
\description{
Plot a customized heatmap with a symmetrical scale that is not stretched by extreme values
}
\author{
Bertram Klinger \email{bertram.klinger@charite.de}
}

