% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_manipulation.R
\name{plotModelAccuracy.MRAmodel}
\alias{plotModelAccuracy.MRAmodel}
\title{Plot model accuracy for MRAmodel}
\usage{
\method{plotModelAccuracy}{MRAmodel}(model_description, limit = Inf,
  show_values = TRUE, graphs = c("accuracy", "diff", "data",
  "simulation", "prediction"), selected_treatments = c(),
  selected_readouts = c(), name = "")
}
\description{
Plot model accuracy for MRAmodel
}
