% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_manipulation.R
\name{plotModelAccuracy.MRAmodelSet}
\alias{plotModelAccuracy.MRAmodelSet}
\title{Plot accuracy of all submodels of a modelset}
\usage{
\method{plotModelAccuracy}{MRAmodelSet}(model_description, limit = Inf,
  show_values = TRUE, graphs = c("accuracy", "data", "simulation"),
  selected_treatments = c(), selected_readouts = c(), name = "")
}
\description{
Plot accuracy of all submodels of a modelset
}
