% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mra_model.R
\name{plotModelGraph}
\alias{plotModelGraph}
\title{Plot the network used in the model}
\usage{
plotModelGraph(mra_model)
}
\arguments{
\item{mra_model}{An MRAmodel object.}
}
\description{
Plot the network used in the model with the experimental design on top of it
}
\author{
Mathurin Dorel \email{dorel@horus.ens.fr}
}

