% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_models.R
\name{plotModelScores.modelGroup}
\alias{plotModelScores.modelGroup}
\title{Plot the measurement scores for a group of models}
\usage{
\method{plotModelScores}{modelGroup}(model_group, ...)
}
\arguments{
\item{model_group}{A modelGroup object}

\item{...}{Passed to base::barplot with a matrix of dimension nr_of_models * nr_of_measurements}
}
\value{
See \code{\link{base::barplot}}
}
\description{
Plot the scores for each measured node fitting as a barplot with a bar for each model for each measurement
}
