% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_manipulation.R
\name{plotModelScores.MRAmodel}
\alias{plotModelScores.MRAmodel}
\title{Plot the scores of each antibody}
\usage{
\method{plotModelScores}{MRAmodel}(mra_model, ...)
}
\arguments{
\item{mra_model}{An MRAmodel object}
}
\description{
Plot the scores of the fit for each antibody, which is how much
of the variation in the data is explained by the model
}
\author{
Mathurin Dorel \email{dorel@horus.ens.fr}
}

