% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_models.R
\name{plotModelScores}
\alias{plotModelScores}
\title{Plot the fitting scores of the model}
\usage{
plotModelScores(x, ...)
}
\description{
Generic function to plot scores for measured nodes fitting
}
\seealso{
plotScores.modelGroup, plotScores.MRAmodel
}

