% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation_model.R
\name{plotModelSimulation}
\alias{plotModelSimulation}
\title{Plot the predictions by the model
One plot per measured node
Plot with error bars if available, and give absolute value}
\usage{
plotModelSimulation(model, targets = "all", readouts = "all",
  log_axis = F)
}
\arguments{
\item{model}{An MRAmodel object}

\item{targets}{A perturbation matrix describing the conditions to be simulated}

\item{readouts}{Vector with the name of the nodes that should be used as readouts}

\item{log_axis}{Boolean, whether the ordinate axis should be in log scale}

\item{maxPlotsPerFrame}{Maximum number of perturbation per frame}
}
\value{
Invisibly, the matrix of the results of the simulation
}
\description{
Plot the predictions by the model
One plot per measured node
Plot with error bars if available, and give absolute value
}
\author{
Mathurin Dorel \email{dorel@horus.ens.fr}
}
\seealso{
getCombinationMatrix
}

