% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_model.R
\name{plotNetworkGraph}
\alias{plotNetworkGraph}
\title{Plot a graph from an adjacency list}
\usage{
plotNetworkGraph(links_list)
}
\arguments{
\item{links_list}{A 2-columns matrix. The network as an adjacency list, the first column is the upstream nodes, the second column the downstream nodes.}
}
\description{
Plot a graph from an adjacency list
}

