% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_models.R
\name{plotParameters.modelGroup}
\alias{plotParameters.modelGroup}
\title{Plot the parameters of the models}
\usage{
\method{plotParameters}{modelGroup}(model_group)
}
\arguments{
\item{model_group}{A modelGroup object}
}
\description{
Plot the parameters fitted for each model as an heatmap
}
\seealso{
Other Model plots: \code{\link{plotModelAccuracy}},
  \code{\link{plotModelGraph}},
  \code{\link{plotModelScores.MRAmodel}},
  \code{\link{plotModelScores}}
}
