% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_models.R
\name{plotParameters}
\alias{plotParameters}
\title{Plot parameters}
\usage{
plotParameters(x, ...)
}
\arguments{
\item{x}{A Model object}

\item{...}{Plot parameters}
}
\description{
Generic function to plot parameters
}
\seealso{
plotParameters.modelGroup
}
