% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compare_models.R
\name{plotScores.modelGroup}
\alias{plotScores.modelGroup}
\title{Plot the scores of each model in the group}
\usage{
\method{plotScores}{modelGroup}(model_group)
}
\arguments{
\item{model_group}{Object of class modelGroup}
}
\value{
Invisibly, the scores ( -log(best_fit/base_fit) )
}
\description{
Display the scores of the modelGroup object as a barplot
}
\seealso{
\code{\link{MRAmodel}}
}

