% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_likelihood.R
\name{exportProfiles}
\alias{exportProfiles}
\alias{importProfiles}
\alias{checkProfiles}
\title{Save profile likelihood results in a .rds file}
\usage{
exportProfiles(profiles, prefix = "likelihood_profiles")

importProfiles(fname)

checkProfiles(profiles)
}
\arguments{
\item{profiles}{A profile likelihood object as produced by profileLikelihood}

\item{prefix}{The prefix for the .rds file.}

\item{fname}{name of the rds file containing}
}
\description{
Save profile likelihood results in a .rds file name "prefix.rds"

Import a .rds file and check that it corresponds to the output of a profile likelihood procedure

Check that an object is a valid profile likelihood
}
