% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_export.R
\name{readMIDAS}
\alias{readMIDAS}
\title{Read a MIDAS file}
\usage{
readMIDAS(fname)
}
\arguments{
\item{fname}{Name of the MIDAS file}
}
\description{
Read a MIDAS file and returns it as a matrix containing the experiments as rows and the readouts as colums
}

