% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_data_generation.R
\name{readNetworkAdj}
\alias{readNetworkAdj}
\title{Get an adjacency matrix from a file}
\usage{
readNetworkAdj(network_file)
}
\arguments{
\item{network_file}{The name of the file containing a network structure as an adjacency matrix, or an adjacency list.}
}
\value{
The weighted adjacency matrix corresponding to the network
}
\description{
Get the adjacency matrix from a various formats of network files
}

