% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model.R
\name{rebuildModelSet}
\alias{rebuildModelSet}
\title{Import a saved set of models with data files to construct a modelSet object}
\usage{
rebuildModelSet(model_files, data_files, var_files = c())
}
\arguments{
\item{model_files}{A list of .mra files containing the information on the models, or a named list of read in mra Files with readLines}

\item{data_files}{A list of .csv files with the data for the models or a named list of read in data files with extractMIDAS}

\item{var_files}{A list of .var files with the variation of the data or a named list of read in var files with extractMIDAS}
}
\value{
An MRAmodelSet object describing the modelSet and its best fit, containing the data
}
\description{
Build a fitted modelSet object from the individual .mra files, and import data for this modelSet
Does NOT perform any initialisation but identifies the variable parameters.
}
\examples{
\dontrun{
rebuildModelSet(c("model1.mra",model2.mra), c("data1.csv","data2.csv"), c("data1.var","data2.var"))
}
}
\seealso{
rebuildModel, createModelSet
}
\author{
Bertram Klinger \email{bertram.klinger@charite.de}
}
