% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_manipulation.R
\name{fitFromModel}
\alias{fitFromModel}
\title{Fit a model using the parameter set from another model}
\usage{
refitModel(mra_model, parameter_set = c(), vary_param = c(),
  inits = 100, nb_cores = 1, method = "randomlhs", fit_name = "")

fitFromModel(mra_model, parameters_model, vary_param = c(),
  inits = 100, nb_cores = 1, method = "randomlhs", fit_name = "")
}
\arguments{
\item{mra_model}{A MRAmodel object}

\item{parameter_set}{A vector of values used as parameters for the model. There must be a many values as there are parameters, or one that will be used for all parameters.}

\item{vary_param}{A vector of index or name of the parameters to refit, the others will be kept constant. Repetitions or redundant information (index and name designating the same parameter) are removed.}

\item{inits}{Number of random initialisations for the variable parameters}

\item{nb_cores}{Number of processes to use for the refitting. 0 to use all cores of the machines but one.}

\item{method}{Method to use for the sample generation for the random initialisations}

\item{fit_name}{Name of the refit for the title of the plots}
}
\value{
The refitted model as an MRAmodel object.
}
\description{
Refit the model with a specified parameter set while keeping parameters constant
Fit a model using the parameter set from another model
}
