% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model.R
\name{refitWithFixedParameter}
\alias{refitWithFixedParameter}
\title{Refit modelset by iterative lumping of variable parameters into a single parameter}
\usage{
refitWithFixedParameter(
  var_par,
  modelset,
  nb_sub_params,
  nb_cores = 0,
  nb_samples = 5,
  method = "geneticlhs"
)
}
\arguments{
\item{var_par}{ID of the variable parameters to test}

\item{modelset}{An MRAmodelSet object}

\item{nb_sub_params}{Number of parameters per submodel}

\item{nb_cores}{Number of cores to use for the fitting}

\item{nb_samples}{Number of samples to generate for the fitting}

\item{method}{Method to use for the fitting}
}
\value{
A list with the fields 'residuals' (fitted residuals), 'removed_var' (=var_par) and 'params' (the fitted parameter sets corresponding to the residuals)
}
\description{
Refit modelset by iterative lumping of variable parameters into a single parameter
}
\seealso{
refitWithVariableParameter
}
