% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model.R
\name{refitWithVariableParameter}
\alias{refitWithVariableParameter}
\title{Refit modelset by iterative relaxation of single parameters}
\usage{
refitWithVariableParameter(var_par, modelset, nb_sub_params, nb_cores = 0,
  nb_samples = 5, method = "geneticlhs")
}
\description{
Refit modelset by iterative relaxation of single parameters
}

