% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{residuals_plot}
\alias{residuals_plot}
\title{Plot residuals vs rank}
\usage{
residuals_plot(residuals, model_name = "default")
}
\arguments{
\item{A}{list of residuals}
}
\description{
Plot all residuals vs rank and best residuals vs rank to assess the quality of the optimisation
}
