% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model.R
\name{sampleWithCorrelation}
\alias{sampleWithCorrelation}
\title{Produces shifted random samples with some parameters inferred from correlations in the measurements}
\usage{
sampleWithCorrelation(model, core, nb_samples, shift = 0, correlated = "",
  sd = 2, perform_plot = F)
}
\arguments{
\item{model}{An MRAmodel object}

\item{core}{A ModelCore list}

\item{nb_samples}{Number of samples to generate}

\item{shift}{A vector giving the shift to be applied to the normal distribution. Must have the same size as the number of parameters}

\item{correlated}{List of parameters extracted with a linear model that should not be randomly initialised}

\item{sd}{Standard deviation of the normal distribution to use for the sampling}

\item{perform_plot}{Whether the linear models should be plotted}
}
\description{
Produces shifted random samples with some parameters inferred from correlations in the measurements
}

