% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_set.R
\name{setVariableParameters}
\alias{setVariableParameters}
\title{Set the variable parameters of a modelset}
\usage{
setVariableParameters(modelset, parameters_ids)
}
\arguments{
\item{modelset}{An MRAmodelSet object}

\item{parameters_ids}{The ids of the parameters to set as variable across submodels}
}
\value{
The modified MRAmodelSet
}
\description{
Set the parameters that can vary across submodels for an MRAmodelSet
}
\author{
Mathurin Dorel \email{dorel@horus.ens.fr}
}
