% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation_model.R
\name{simulateModel}
\alias{simulateModel}
\title{Get the target simulations in a MIDAS design-like or list format
Simulate a model for a specified set of perturbation}
\usage{
simulateModel(model_description, targets = "all", readouts = "all")
}
\arguments{
\item{model_description}{A MRAmodel object that describes the model to be fitted}

\item{targets}{A matrix of perturbations to simulate, as generated by \link{getCombinationMatrix} or "all" to reproduce the perturbations used to build the model
A perturbation matrix is filled with 0 and 1, each row is a perturbation set and column names are used to determine the name of the perturbation.}

\item{readouts}{List of nodes to simulate. If "all", all the nodes measured to fit the model will be used. Only nodes actually measured or inhibited for the model can be simulated.}
}
\value{
A list that represents a MIDAS measure-like format with fields 'conditions' which is the matrix of perturbations provided as 'targets', 'bestfit' which is the simulation, and 'variants' which is a list of simulations for the alternative parameter sets from profile likelihood
}
\description{
Get the target simulations in a MIDAS design-like or list format
Simulate a model for a specified set of perturbation
}
\seealso{
\code{\link{getCombinationMatrix}}
}

