% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{sub_data_frame}
\alias{sub_data_frame}
\title{Return a data.frame even if the output as on line or one column}
\usage{
sub_data_frame(input, rows = NULL, cols = NULL)
}
\description{
Return a data.frame even if the output as on line or one column
}
