% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_manipulation.R
\name{suggestExtension}
\alias{suggestExtension}
\title{Tries to add one link each and returns a list of links ordered by their chi-squared differences to the original model
This list can then be based and compared to literature knowledge and if considered suitable manually added to the starting network and rerun with the createModel function.}
\usage{
suggestExtension(model_description, parallel = F, mc = 1,
  sample_range = c(10^(2:-1), 0, -10^(-1:2)), print = F)
}
\arguments{
\item{model_description}{MRAmodel object describing thze model and its best fit, containing the data}

\item{parallel}{Boolean number indicating whether addition is executed in a parallel fashion}

\item{mc}{Number of cores that should be used for the computation}

\item{sample_range}{Numeric vector containing all starting values for the new link (DEFAULT: c(10^(2:-1),0,-10^(-1:2)))}

\item{print}{Boolean indicating whether the result should be printed in a text file "Additional_link_suggestion.txt"}
}
\description{
Tries to add one link each and returns a list of links ordered by their chi-squared differences to the original model
This list can then be based and compared to literature knowledge and if considered suitable manually added to the starting network and rerun with the createModel function.
}
\examples{
\dontrun{
ext_list = suggestExtension(mramodel)
}
}
\author{
Bertram Klinger \email{bertram.klinger@charite.de}
}
\seealso{
selectMinimalModel, createModel
}

