% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_manipulation.R
\name{addLink}
\alias{addLink}
\title{add Link routine}
\usage{
addLink(
  new_link,
  adj,
  rank,
  init_residual,
  model,
  initial_response,
  expdes,
  data,
  model_structure,
  sample_range,
  variable_links = c(),
  verbose = F
)
}
\arguments{
\item{new_link}{Integer link whose addition is to be tested}

\item{adj}{integer Matrix original adjacency matrix excluding the new_link}

\item{rank}{Rank of the input model}

\item{init_residual}{Numeric sum-squared error of original network}

\item{model}{MRAmodel or MRAmodelSet object of original network}

\item{initial_response}{List containing the local_response matrix and inhibitor strength of original network}

\item{expdes}{Design object of MRAmodel object}

\item{data}{data Object of MRAmodel object}

\item{model_structure}{Structure object of MRAmodel object}

\item{sample_range}{Numeric vector containing all starting values for new_link}

\item{verbose}{Whether the function should be verbose or not}
}
\description{
add Link routine
}
\author{
Bertram Klinger \email{bertram.klinger@charite.de}
}
