% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_likelihood.R
\name{addPLinfos}
\alias{addPLinfos}
\title{Add the information provided by the profile likelihood in the model_description object
ie the limits of the confidence interval and the corresponding sets of parameters}
\usage{
addPLinfos(model_description, profiles)
}
\arguments{
\item{model_description}{An MRAmodel object}

\item{profiles}{A list of the likelihood profiles of the parameters of the model}
}
\value{
An MRAmodel object with the profile likelihood information integrated to it
}
\description{
Add the information provided by the profile likelihood in the model_description object
ie the limits of the confidence interval and the corresponding sets of parameters
}
\seealso{
\code{\link{profileLikelihood}}
}
\author{
Mathurin Dorel \email{dorel@horus.ens.fr}
}
