% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model.R
\name{addVariableParameters}
\alias{addVariableParameters}
\title{Search for and relax fixed coefficients of an MRAmodelSet to fit individual data sets significantly better}
\usage{
addVariableParameters(original_modelset, nb_cores = 0,
  max_iterations = 0, nb_samples = 100, accuracy = 0.95,
  method = "geneticlhs", notVariable = c(), max_variables = Inf)
}
\arguments{
\item{original_modelset}{An MRAmodelSet object}

\item{nb_cores}{Number of cores to use for the refitting with the new variable coefficients}

\item{max_iterations}{Maximum number of variable parameters to add (if 0 takes as many as possible)}

\item{nb_samples}{Number of samples to generate to fit the new variable parameters}

\item{accuracy}{Cutoff probability for the chi^2 test}

\item{method}{Name of the LHS method to use}

\item{notVariable}{Coefficients that should not be varied, either a numeric or character vector identifying the coefficients from 'modelset$model$getParametersNames()$names', defaults to 'c()'.}

\item{max_variables}{Maximum number of parameters to vary}
}
\value{
An updated MRAmodelSet with the new set of coefficients
}
\description{
See if an MRAmodelSet requires some coefficients to be variable among models to explain the variations
}
\author{
Mathurin Dorel \email{dorel@horus.ens.fr}

Bertram Klinger \email{bertram.klinger@charite.de}
}
