% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model_set.R
\name{compareModels}
\alias{compareModels}
\title{Compare networks}
\usage{
compareModels(files)
}
\arguments{
\item{files}{A list of .mra files}
}
\value{
None
}
\description{
Compare the parameters of one network for different conditions (cell line, perturbations, ...)
The links must be the same for all the networks
}

