% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_set.R
\name{compareParameters}
\alias{compareParameters}
\alias{compareParameters.MRAmodelSet}
\alias{compareParameters.modelGroup}
\title{Compare networks}
\usage{
compareParameters(multi_model, ...)

\method{compareParameters}{MRAmodelSet}(modelset)

\method{compareParameters}{modelGroup}(modelgroup)
}
\arguments{
\item{modelset}{A MRAmodelSet object}
}
\value{
matrix of parameters
}
\description{
Compile the parameters of a model set in one matrix and plot the results.
highlights the variable links by scaling to it
}
\author{
Bertram Klinger \email{bertram.klinger@charite.de}

Mathurin Dorel \email{dorel@horus.ens.fr}
}
