% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mra_model.R
\name{computeFitScore}
\alias{computeFitScore}
\title{Compute the fitting scores of a model}
\usage{
computeFitScore(mra_model, refit_model = F)
}
\arguments{
\item{mra_model}{The MRAmodel object for which the score should be computed}

\item{refit_model}{Whether the model should be refitted before computing the scores}
}
\value{
A MRAmodel object with the scores in the fields 'Rscores' and 'bestfitscore'
}
\description{
Compute 2 fitting scores for the model, the fraction of the variation in the data explained by the network, and the improvement compared to a model with no links
Do the computation for each measured node and for the network
}

