% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_model.R
\name{createModel}
\alias{createModel}
\title{Creates a parameterised model from experiment files and the network structure, and fit the parameters to the data}
\usage{
createModel(model_links, data.stimulation, basal_file, data.variation = "",
  nb_cores = 1, inits = 1000, init_distribution = F, method = "default")
}
\arguments{
\item{model_links}{Path to the file containing the network structure, either in matrix form or in list of links form. Extension .tab expected}

\item{data.stimulation}{Path to the file containing the data in MRA_MIDAS format. Extension .csv expected.}

\item{basal_file}{Path to the file indicating the nodes without basal activity. Extension .dat expected.}

\item{data.variation}{Path to the file containing the coefficient of variation for each measurement in MRA_MIDAS format. If it is not provided, the function uses the replicates of the data.stimulation file to determine a variance per probe (i.e antibody/DNA fragment/...). Extension .var expected.}

\item{nb_cores}{Number of cores that should be used for the computation}

\item{inits}{Number of initialisation steps which should be performed (see method for the exact meaning of this value)}

\item{init_distribution}{Whether the distribution of the residuals and the parameters deduced by correlation should be plotted or not}

\item{method}{Method to be used for the initialisation, available methods are :
random : Perform a Latin Hypercube Sampling to choose \emph{inits} starting points then perform a gradient descent to find the local optimum for each of those points.
correlation : Deduce some parameters from the correlation between the measurements for the target node and all of its input nodes, then perform random to find the other parameters. Recommended, very efficient for small datasets.
genetic : Genetic algorithm with mutation only. \emph{inits} is the total number of points sampled.
annealing : Simulated annealing and gradient descent on the best result. \emph{inits} is the maximum number of iteration without any change before the algorithm decides it reached the best value. Use not recommended.}
}
\value{
An MRAmodel object describing the model and its best fit, containing the data
}
\description{
Creates a parameterised model from experiment files and the network structure, and fit the parameters to the data
}
\examples{
model = createModel("links.tab", "data_MIDAS.csv", "basal.dat") # Produces a model for the network described in links.tab using the data in data_MIDES.csv
model = createModel("links.tab", "data_MIDAS.csv", "basal.dat", "variation.var") # Uses the variation from a variation file
model = createModel("links.tab", "data_MIDAS.csv", "basal.dat", nb_cores = detectCores()) # Uses all cores available (with the package parallel)
model = createModel("links.tab", "data_MIDAS.csv", "basal.dat", inits = 1000000) # Uses more initialisations for a complex network
}
\author{
Mathurin Dorel \email{dorel@horus.ens.fr}
}
\seealso{
importModel, exportModel, rebuildModel
}

