% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model.R
\name{extractModelCore}
\alias{extractModelCore}
\title{Extracts the data, the experimental design and the structure from the input files
links must be a matrix of links [node1, node2]
Experiment file dta_file syntax should be as follows, with one line per replicate
         stimulator                |          inhibitor                |                         type                       | [one column per measured nodes]
-------------------------------------------------------------------------------------------------------------------------------------------------------------
stimulator name or solvant if none | inhibitor name or solvant if none | c for control, b for blank and t for experiment |    measure for the condition}
\usage{
extractModelCore(model_structure, basal_activity, data_filename,
  var_file = "")
}
\description{
Extracts the data, the experimental design and the structure from the input files
links must be a matrix of links [node1, node2]
Experiment file dta_file syntax should be as follows, with one line per replicate
         stimulator                |          inhibitor                |                         type                       | [one column per measured nodes]
-------------------------------------------------------------------------------------------------------------------------------------------------------------
stimulator name or solvant if none | inhibitor name or solvant if none | c for control, b for blank and t for experiment |    measure for the condition
}

