% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model.R
\name{extractModelCore}
\alias{extractModelCore}
\title{Extracts the data, the experimental design and the structure from the input files}
\usage{
extractModelCore(model_structure, basal_activity, datas, var_file = "",
  dont_perturb = c(), dont_read = c(), MIN_CV = 0.1, DEFAULT_CV = 0.3)
}
\arguments{
\item{model_structure}{Matrix of links [node1, node2]}

\item{basal_activity}{The node of the structure with a basal activity}

\item{datas}{Experimental data under the MIDAS format. See extractMIDAS.}

\item{var_file}{Variation file name under MIDAS format or "" to compute an error from the data. See extractMIDAS.}

\item{dont_perturb}{Perturbations to be removed for the fit, will not be used nor simulated. (vector of names)}

\item{dont_read}{Readouts to be removed for the fit, will not be used nor simulated. (vector of names)}

\item{MIN_CV}{Minimum coefficient of variation.}

\item{DEFAULT_CV}{Default coefficient of variation to use when none is provided and there are no replicated in the data.}
}
\description{
Extracts the data, the experimental design and the structure from the input files
}
\seealso{
\code{\link{extractMIDAS}}
}

