% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model.R
\name{extractModelCore}
\alias{extractModelCore}
\title{Extracts the data, the experimental design and the structure from the input files}
\usage{
extractModelCore(model_structure, basal_activity, data_filename,
  var_filename = "", dont_perturb = c(), dont_read = c(),
  MIN_CV = 0.1, DEFAULT_CV = 0.3, data_space = "log")
}
\arguments{
\item{model_structure}{Matrix of links [node1, node2]}

\item{basal_activity}{The node of the structure with a basal activity}

\item{dont_perturb}{Perturbations to be removed for the fit, will not be used nor simulated. (vector of names)}

\item{dont_read}{Readouts to be removed for the fit, will not be used nor simulated. (vector of names)}

\item{MIN_CV}{Minimum coefficient of variation.}

\item{DEFAULT_CV}{Default coefficient of variation to use when none is provided and there are no replicates in the data.}

\item{data_space}{One of "log" or "linear". Determines whether the data should be fitted in log space or linear space. The parameters will be in log space in both cases, but the log space data might be more adapted if the data are log-normal (as opposed to normal assumed by the linear option).}

\item{datas}{Experimental data under the MIDAS format. See extractMIDAS.}

\item{var_file}{Variation file name under MIDAS format or "" to compute an error from the data. See extractMIDAS.}
}
\description{
Extracts the data, the experimental design and the structure from the input files
}
\seealso{
\code{\link{extractMIDAS}}
}
