% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitmodel.R
\docType{package}
\name{fitmodel}
\alias{fitmodel}
\alias{fitmodel-package}
\title{Fit and simulate model from perturbation data}
\description{
Fits and simulates network model derived from perturbation data using Modular Response Analysis and Profile Likelihood
}
\author{
Bertram Klinger \email{bertram.klinger@charite.de}, Mathurin Dorel \email{dorel@horus.ens.fr}, Nils Blüthgen \email{nils.bluethgen@charite.de}
}

