% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_data_generation.R
\name{generateToyNetwork}
\alias{generateToyNetwork}
\title{Generate a network topology for testing and benchmarking}
\usage{
generateToyNetwork(
  ninputs,
  p_link = 0.2,
  layers = c(3, 3),
  nfeedbacks = 0,
  p_layer_link = 0,
  fname = ""
)
}
\arguments{
\item{ninputs}{An integer, the number of inputs in the model}

\item{p_link}{Number or vector of number between 0 and 1 (length=length(layers)). Probability of linking of each node in a layer with each node in the upper layer.}

\item{layers}{A vector of integers, The length of the vector is the number of layers besides the receptor layer, and the integers are the number of nodes in each layer}

\item{nfeedbacks}{An integer, the number of feedback (link from a bottom layer to an upper layer)}

\item{p_layer_link}{Number or vector of number. Probability of two nodes in a layer being directly linked}

\item{fname}{A filename where the adjacency list should be saved, "" for no saving.}
}
\value{
A 2-columns matrix representing the network as an adjacency list
}
\description{
Generate a network topology for testing and benchmarking
}
\seealso{
Other toy data generation: 
\code{\link{addLayerLinks}()},
\code{\link{createSimulation}()},
\code{\link{generateToyDesign}()}
}
\concept{toy data generation}
