% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_comparison.R
\name{printDirectPaths}
\alias{printDirectPaths}
\title{Print the direct paths with confidence interval}
\usage{
printDirectPaths(mra_model, digits = 2)
}
\arguments{
\item{digits}{Number of digits to display for the confidence interval values}
}
\value{
Hidden. The direct paths
}
\description{
Print the direct paths with confidence interval
}
\seealso{
getDirectPaths
}
