% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model.R
\name{getSamples}
\alias{getSamples}
\title{Get LHS samples}
\usage{
getSamples(sample_size, nb_samples, method = "randomlhs")
}
\arguments{
\item{sample_size}{Size of each sample}

\item{nb_samples}{Number of samples}

\item{method}{Name of the LHS method to use}
}
\description{
Get LHS samples
}

