% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_model.R
\name{getSimulation}
\alias{getSimulation}
\title{Simulate the model for the experimental design used for the fitting}
\usage{
getSimulation(mra_model, with_offset = TRUE)
}
\arguments{
\item{mra_model}{The MRAmodel to simulate}
}
\value{
A matrix containing the simulation with the names of the measured nodes as column names
}
\description{
Simulate the model for the experimental design used for the fitting
}
