% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_export.R
\name{importModel}
\alias{importModel}
\title{Import model from a file}
\usage{
importModel(file_name = NULL, file = NULL)
}
\arguments{
\item{file_name}{Name of the .mra file}

\item{file}{R object from an .mra file that has been read in by readLines}
}
\value{
An MRAmodel object
}
\description{
Import an MRAmodel object from a .mra file or from an .mra file that has been read in with readLines.
}
\seealso{
exportModel, rebuildModel
}
\author{
Mathurin Dorel \email{dorel@horus.ens.fr}
}
