% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_model.R
\name{initModel}
\alias{initModel}
\title{Perform an initialisation of the model
Possibility to use different sampling methods}
\usage{
initModel(model, core, inits, precorrelate = T, method = "randomlhs",
  nb_cores = 1, perform_plots = F)
}
\description{
Perform an initialisation of the model
Possibility to use different sampling methods
}

