% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_set.R
\name{MRAmodelSet}
\alias{MRAmodelSet}
\title{Constructor for MRAmodelSet objects}
\usage{
MRAmodelSet(
  nb_models = 1,
  model = NULL,
  design = NULL,
  structure = NULL,
  basal = matrix(),
  data = matrix(),
  cv = matrix(),
  parameters = vector(),
  bestfit = NA,
  name = c(),
  infos = c(),
  param_range = list(),
  lower_values = c(),
  upper_values = c(),
  unused_perturbations = c(),
  unused_readouts = c(),
  min_cv = 0.1,
  default_cv = 0.3,
  use_log = FALSE
)
}
\arguments{
\item{nb_models}{Number of models in the set}

\item{model}{An object of class Model}

\item{design}{An object of class ExperimentalDesign}

\item{structure}{An object of class ModelStructure}

\item{basal}{A matrix describing the basal activity of the nodes in the network}

\item{data}{An object of class Data}

\item{cv}{A matrix containing the coefficient of variation of the data used to build the model}

\item{parameters}{A vector containing the values of the parameters of the model}

\item{bestfit}{The residual chi-2 value associated with the best fit}

\item{name}{Name of the model}

\item{infos}{Extra information on the model}

\item{param_range}{Alternative parameters sets for the model}

\item{lower_values}{Lower bound on the values of the parameters}

\item{upper_values}{Upper bound on the values of the parameters}

\item{unused_perturbations}{Perturbations from the data file that have not been used for the fitting}

\item{unused_readouts}{Readouts from the data file that have not been used for the fitting}
}
\value{
An MRAmodelSet object
}
\description{
Build an MRAmodelSet, which contains all the information required to simulate a set of MRAmodels, ensuring that the sets of parameters are similar for all models
}
\seealso{
\code{\link{createModel}}
}
\author{
Mathurin Dorel \email{dorel@horus.ens.fr}
}
