% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model_manipulation.R
\name{plotModelAccuracy}
\alias{plotModelAccuracy}
\title{Plot heatmaps of the model simulation against the data weighted by the error, as well as the log fold change for the data and the prediction}
\usage{
plotModelAccuracy(model_description)
}
\arguments{
\item{model_description}{An MRAmodel object}
}
\value{
Nothing
}
\description{
Plot heatmaps of the model simulation against the data weighted by the error, as well as the log fold change for the data and the prediction
}
\author{
Mathurin Dorel \email{dorel@horus.ens.fr}
}
\seealso{
plotModelSimulation, createModel, importModel
}

