% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_manipulation.R
\name{plotModelAccuracy}
\alias{plotModelAccuracy}
\title{Plot heatmaps of the model simulation against the data weighted by the error, as well as the log fold change for the data and the prediction}
\usage{
plotModelAccuracy(model_description, limit = Inf, show_values = TRUE)
}
\arguments{
\item{model_description}{An MRAmodel object}

\item{limit}{An integer to force the limit of the heatmaps}

\item{show_values}{Whether the values should be printed in the heatmap boxes or not.}
}
\value{
Nothing
}
\description{
Plot heatmaps of the model simulation against the data weighted by the error, as well as the log fold change for the data and the prediction
}
\seealso{
createModel, importModel

Other Model plots: \code{\link{plotModelGraph}},
  \code{\link{plotModelScores.MRAmodel}},
  \code{\link{plotModelScores}},
  \code{\link{plotParameters.modelGroup}}
}
\author{
Mathurin Dorel \email{dorel@horus.ens.fr}
}
\concept{Model plots}
