% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mra_model.R
\name{plotModelGraph}
\alias{plotModelGraph}
\title{Plot the network used in the model}
\usage{
plotModelGraph(mra_model)
}
\arguments{
\item{mra_model}{A MRAmodel object.}
}
\description{
Plot the network used in the model with the experimental design on top of it
}
\seealso{
Other Model plots: \code{\link{plotModelAccuracy}},
  \code{\link{plotModelScores.MRAmodel}},
  \code{\link{plotModelScores}},
  \code{\link{plotParameters.modelGroup}}

Other Network graph: \code{\link{plotNetworkGraph}}
}
\author{
Mathurin Dorel \email{dorel@horus.ens.fr}
}
