% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mra_model.R
\name{plotModelGraph}
\alias{plotModelGraph}
\title{Plot the network used in the model}
\usage{
plotModelGraph(mra_model, print_values = TRUE, scaling = 5,
  max_width = 3, min_width = 1)
}
\arguments{
\item{mra_model}{A MRAmodel object.}

\item{print_values}{Whether the parameter values should be printed on top of the arrows in the graph.}
}
\description{
Plot the network used in the model with the experimental design on top of it
}
\seealso{
Other Model plots: \code{\link{plotModelAccuracy}},
  \code{\link{plotModelScores.MRAmodel}},
  \code{\link{plotModelScores}},
  \code{\link{plotParameters}}, \code{\link{plotResiduals}}

Other Network graph: \code{\link{plotNetworkGraph}}
}
\author{
Mathurin Dorel \email{dorel@horus.ens.fr}
}
\concept{Model plots}
\concept{Network graph}
