% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_comparison.R
\name{plotModelParameters}
\alias{plotModelParameters}
\title{Plot the parameters for a model or a list of models}
\usage{
plotModelParameters(models, non_stop_nodes = c(), lim = 2)
}
\arguments{
\item{non_stop_nodes}{A list of nodes that should not begin or end a path, usefull to compare models with different basal activities. If the node is only present as a first or last node of a path, the corresponding paths will be deleted.}

\item{lim}{The absolute limit of the plot y-axis, the parameters with value falling outside the range [-lim, lim] will have their value displayed at the edge of the plotting area.}

\item{aggregated_paths}{models A list of models where each entry must be named (i.e. 'names(models)' must return a vector)}

\item{digits}{Number of digits to display for the confidence interval values}
}
\value{
Invisibly the aggregated direct paths
}
\description{
Plot the parameters for a model or a list of models
}
\examples{
\donotrun{ plotModelParameters }
}
\seealso{
\code{\link{plotParameters}}, \code{\link{getDirectPaths}}, \code{\link{aggregateDirectPaths}}
}
\author{
Mathurin Dorel \email{dorel@horus.ens.fr}
}
