% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_manipulation.R
\name{plotModelScores.MRAmodel}
\alias{plotModelScores.MRAmodel}
\title{Plot the scores of each antibody}
\usage{
\method{plotModelScores}{MRAmodel}(mra_model, ...)
}
\arguments{
\item{mra_model}{An MRAmodel object}

\item{...}{Extra barplot parameters}
}
\description{
Plot the scores of the fit for each antibody, which is how much
of the variation in the data is explained by the model
}
\seealso{
Other Model plots: \code{\link{plotModelAccuracy}},
  \code{\link{plotModelGraph}},
  \code{\link{plotModelScores}},
  \code{\link{plotParameters}}, \code{\link{plotResiduals}}
}
\author{
Mathurin Dorel \email{dorel@horus.ens.fr}
}
\concept{Model plots}
