% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_model.R
\name{plotModelSimulation}
\alias{plotModelSimulation}
\title{Plot the predictions by the model
One plot per measured node
Plot with error bars if available, and give absolute value}
\usage{
plotModelSimulation(prediction, log_axis = F)
}
\arguments{
\item{prediction}{A list of the model predictions as produced by \link{simulateModel}}

\item{log_axis}{Boolean, whether the ordinate axis should be in log scale}

\item{maxPlotsPerFrame}{Maximum number of perturbation per frame}
}
\value{
Invisibly, the matrix of the results of the simulation
}
\description{
Plot the predictions by the model
One plot per measured node
Plot with error bars if available, and give absolute value
}
\author{
Mathurin Dorel \email{dorel@horus.ens.fr}
}
\seealso{
getCombinationMatrix, simulateModel
}

