% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_model.R
\name{plotModelSimulation}
\alias{plotModelSimulation}
\title{Plot the predictions by the model}
\usage{
plotModelSimulation(
  model_description,
  targets = "all",
  readouts = "all",
  inhibition_effect = NA,
  log_axis = TRUE,
  with_data = FALSE,
  compare = list()
)
}
\arguments{
\item{model_description}{A MRAmodel object that describes the model to be fitted}

\item{targets}{A matrix of perturbations to simulate, as generated by \link{getCombinationMatrix} or "all" to reproduce the perturbations used to build the model
A perturbation matrix is filled with 0 and 1, each row is a perturbation set and column names are used to determine the name of the perturbation.}

\item{readouts}{List of nodes to simulate. If "all", all the nodes measured to fit the model will be used. Only nodes actually measured or inhibited for the model can be simulated.}

\item{inhibition_effect}{A single value, a list of values or NA. Values in ]0, -inf] to use for the inhibition, representing the log2-fold change in activity of the node (alternatively, a value between 0 and 1 representing the fraction of activity remaining after inhibition compared to basal). If NA, the values fitted for the inhibition will be used, or -1 if an inhibition is requested for a node that was not inhibited in the experiment.}

\item{log_axis}{Boolean, whether the ordinate axis should be in log scale}

\item{with_data}{Plot the data of the model next to the prediction}

\item{compare}{A list of MRAmodel to compare the predictions, targets and readouts must be valid for MRAmodels}
}
\value{
A list that represents a MIDAS measure-like format with fields 'conditions' the matrix of perturbations provided as 'targets', 'bestfit' the simulation, and 'variants' a list of simulations for the alternative parameter sets from profile likelihood

Invisibly, the matrix of the results of the simulation
}
\description{
Plot the predictions by the model
}
\author{
Mathurin Dorel \email{mathurin.dorel@charite.de}
}
