% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model.R
\name{plotNetworkGraph}
\alias{plotNetworkGraph}
\title{Plot a graph from an adjacency list}
\usage{
plotNetworkGraph(structure, expdes = "", local_values = "")
}
\arguments{
\item{structure}{A 2-columns matrix or a ModelStructure object. The network as an adjacency list, the first column is the upstream nodes, the second column the downstream nodes. Or a ModelStructure object as returned by getModelStructure.}

\item{expdes}{An ExperimentalDesign object. The measured, stimulated and inhibited nodes are highlighted if present. Signalling strengths are indicated in leftshifted edges and inhibitor strengths are denoted in red below the inhibited node.}

\item{local_values}{A list with entries 'local_response' (A weighted adjacency matrix representing the values of the links) and 'inhibitors' (A list of inhibition values) both compatible with the 'structure' input}
}
\description{
Plot a graph from an adjacency list
}
\seealso{
Other Network graph: \code{\link{plotModelGraph}}
}
\author{
Bertram Klinger \email{bertram.klinger@charite.de}
}
