% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model.R
\name{plotNetworkGraph}
\alias{plotNetworkGraph}
\title{Plot a graph from an adjacency list}
\usage{
plotNetworkGraph(links_list, expdes = "", local_values = "",
  edge_lim = c(1, 10, 1))
}
\arguments{
\item{links_list}{A 2-columns matrix or a ModelStructure object. The network as an adjacency list, the first column is the upstream nodes, the second column the downstream nodes. Or a ModelStructure object as returned by getModelStructure.}

\item{expdes}{An ExperimentalDesign object. The measured, stimulated and inhibited nodes are highlighted if present.}
}
\description{
Plot a graph from an adjacency list
}

