% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_models.R, R/parameters_comparison.R
\name{plotParameters}
\alias{plotParameters}
\title{Plot parameters}
\usage{
plotParameters(aggregated_paths, lim = 2)

plotParameters(aggregated_paths, lim = 2)
}
\arguments{
\item{aggregated_paths}{Aggregated direct paths as returned by aggregateDirectPaths}

\item{lim}{The absolute limit of the plot y-axis, the parameters with value falling outside the range [-lim, lim] will have their value displayed at the edge of the plotting area.}

\item{x}{A Model object}

\item{...}{Plot parameters}
}
\description{
Generic function to plot parameters

Plot parameters from aggregated direct paths. The margins are increased to c(10, 4, 4, 2) but can be increased more by specifying larger margins in 'par' before calling the function.
}
\seealso{
plotParameters.modelGroup

\code{\link{aggregateDirectPaths}}
}

