% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_models.R
\name{plotResiduals}
\alias{plotResiduals}
\title{Plot the residual of each model in the group}
\usage{
plotResiduals(model_group, main = "Model residuals")
}
\arguments{
\item{model_group}{Object of class modelGroup}
}
\value{
Invisibly, the residuals (chi-2)
}
\description{
Display the residuals of the models as a barplot
}
\seealso{
Other Model plots: \code{\link{plotModelAccuracy}},
  \code{\link{plotModelGraph}},
  \code{\link{plotModelScores.MRAmodel}},
  \code{\link{plotModelScores}},
  \code{\link{plotParameters}}
}
\author{
Bertram Klinger \email{bertram.klinger@charite.de}
}
\concept{Model plots}
