% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_model.R
\name{plotSimulation}
\alias{plotSimulation}
\title{Plot predictions generated by simulateModel
One plot per measured node
Plot with error bars if available, and give absolute value}
\usage{
plotSimulation(
  prediction,
  log_axis = FALSE,
  with_data = TRUE,
  data_color = cbbPalette[1],
  sim_colors = cbbPalette[-1],
  compare = list(),
  strict = TRUE
)
}
\arguments{
\item{prediction}{A list of the model predictions as produced by \link{simulateModel}}

\item{log_axis}{Boolean, whether the ordinate axis should be in log scale}

\item{with_data}{Display 2 bars per condition, one for the simulation and one for the data}

\item{data_color}{Color of the bars corresponding to the data}

\item{sim_colors}{Vector of colors for the bars corresponding to the simulations. The first color is used for the main prediction and the next for the predictions in 'compare', recycled as necessary for all predictions.}

\item{compare}{A list of predictions to compare to. A predition is a list with fields 'bestfit', 'conditions', 'error', 'data', 'unstim_data' and 'variants' as generated by simulateModel}

\item{strict}{When 'compare' is not empty, whether the other prediction conformity must be strictly evaluated (presence and equality of the field 'condition', equal name of columns in field 'bestfit') or not (the 'bestfit' length is the only thing that is checked)}
}
\value{
Invisibly, the matrix of the results of the simulation
}
\description{
Plot predictions generated by simulateModel
One plot per measured node
Plot with error bars if available, and give absolute value
}
\seealso{
getCombinationMatrix, simulateModel

Other simulation: 
\code{\link{getCombinationMatrix}()},
\code{\link{simulateModel}()}
}
\author{
Mathurin Dorel \email{mathurin.dorel@charite.de}
}
\concept{simulation}
