% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_manipulation.R
\name{printParameters}
\alias{printParameters}
\title{Print the value of each path from the model, with the profile likelihood infos if they are provided}
\usage{
printParameters(model_description, precision = 2)
}
\arguments{
\item{model_description}{An MRAmodel object}

\item{precision}{Number of significant digits to print}
}
\value{
Nothing
}
\description{
Print the value of each path from the model, with the profile likelihood infos if they are provided
}
\author{
Mathurin Dorel \email{dorel@horus.ens.fr}
}

