% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model.R
\name{rebuildModel}
\alias{rebuildModel}
\title{Import a save model with data files}
\usage{
rebuildModel(model_file, data_file, var_file = "")
}
\arguments{
\item{model_file}{A .mra file containing the information on the model}

\item{data_file}{A .csv file with the data for the model}

\item{var_file}{A .var file with the variation of the data}
}
\value{
An MRAmodel object describing the model and its best fit, containing the data
}
\description{
Build a fitted model from a .mra file, and import data for this model
Does NOT perform any initialisation
}
\examples{
\dontrun{
rebuildModel("model.mra", "data.csv", "data.var")
}
}
\author{
Mathurin Dorel \email{dorel@horus.ens.fr}
}
\seealso{
importModel, exportModel, createModel
}

