% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_model.R
\name{sampleWithCorrelation}
\alias{sampleWithCorrelation}
\title{Produces shifted random samples with some parameters inferred from correlations in the measurements}
\usage{
sampleWithCorrelation(model, core, nb_samples, shift = 0, correlated = "",
  sd = 2, plot = F)
}
\arguments{
\item{model}{An MRAmodel object}

\item{shift}{A vector giving the shift to be applied to the normal distribution. Must have the same size as the number of parameters}
}
\description{
Produces shifted random samples with some parameters inferred from correlations in the measurements
}

