% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_model.R
\name{simulateModel}
\alias{simulateModel}
\title{Get the target simulations in a MIDAS design-like or list format
Simulate a model for a specified set of perturbation}
\usage{
simulateModel(model_description, targets, readouts = "all")
}
\arguments{
\item{model_description}{A MRAmodel object that describes the model to be fitted}

\item{targets}{A matrix of perturbations to simulate, as generated by \link{getCombinationMatrix}. This is a matrix filled with 0 and 1, each row is a perturbation set, column names are used to determine the name of the perturbation.}

\item{readouts}{List of nodes to simulate. If "all", all the nodes measured to fit the model will be used. Only nodes actually measured or inhibited for the model can be simulated.}
}
\value{
A matrix in a MIDAS measure-like format
}
\description{
Get the target simulations in a MIDAS design-like or list format
Simulate a model for a specified set of perturbation
}
\seealso{
\code{\link{getCombinationMatrix}}
}

