% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_model.R
\name{simulateModel}
\alias{simulateModel}
\title{Get the target simulations in a MIDAS design-like or list format
Simulate a model for a specified set of perturbation}
\usage{
simulateModel(
  model_description,
  targets = "all",
  readouts = "all",
  inhibition_effect = NA,
  with_offset = TRUE
)
}
\arguments{
\item{model_description}{A MRAmodel object that describes the model to be fitted}

\item{targets}{A matrix of perturbations to simulate, as generated by \link{getCombinationMatrix} or "all" to reproduce the perturbations used to build the model
A perturbation matrix is filled with 0 and 1, each row is a perturbation set and column names are used to determine the name of the perturbation.}

\item{readouts}{List of nodes to simulate. If "all", all the nodes measured to fit the model will be used. Only nodes actually measured or inhibited for the model can be simulated.}

\item{inhibition_effect}{A single value, a list of values or NA. Values in ]0, -inf] to use for the inhibition, representing the log2-fold change in activity of the node (alternatively, a value between 0 and 1 representing the fraction of activity remaining after inhibition compared to basal). If NA, the values fitted for the inhibition will be used, or -1 if an inhibition is requested for a node that was not inhibited in the experiment.}

\item{with_offset}{Whether the simulation should include the offset (fitted simulation) or not (real activity prediction)}
}
\value{
A list that represents a MIDAS measure-like format with fields 'conditions' the matrix of perturbations provided as 'targets', 'bestfit' the simulation, and 'variants' a list of simulations for the alternative parameter sets from profile likelihood
}
\description{
Get the target simulations in a MIDAS design-like or list format
Simulate a model for a specified set of perturbation
}
\seealso{
Other simulation: 
\code{\link{getCombinationMatrix}()},
\code{\link{plotSimulation}()}
}
\concept{simulation}
