% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_manipulation.R
\name{suggestExtension}
\alias{suggestExtension}
\title{Tries to add one link each and returns a list of ordered chi-squared differences, highlighting the significant ones}
\usage{
suggestExtension(model_description, nb_cores = 1, inits = 1000,
  method = "geneticlhs")
}
\arguments{
\item{model_description}{MRAmodel object describing thze model and its best fit, containing the data}

\item{nb_cores}{Number of cores that should be used for the computation}

\item{inits}{Number of initialisation steps which should be performed (see method for the exact meaning of this value)}

\item{method}{Method to be used for the initialisation, available methods are :
random : Perform a Latin Hypercube Sampling to choose \emph{inits} starting points then perform a gradient descent to find the local optimum for each of those points.
correlation : Deduce some parameters from the correlation between the measurements for the target node and all of its input nodes, then perform random to find the other parameters. Recommended, very efficient for small datasets.
genetic : Genetic algorithm with mutation only. \emph{inits} is the total number of points sampled.
annealing : Simulated annealing and gradient descent on the best result. \emph{inits} is the maximum number of iteration without any change before the algorithm decides it reached the best value. Use not recommended.}
}
\value{
An MRAmodel object describing the model and its best fit, containing the data
}
\description{
Tries to add one link each and returns a list of ordered chi-squared differences, highlighting the significant ones
}
\examples{
ext_list = createModel(model) # Produces a model for the network described in links.tab using the data in data_MIDES.csv
}
\author{
Bertram Klinger \email{bertram.klinger@charite.de}
}
\seealso{
createModel, initModel
}

