% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_manipulation.R
\name{suggestExtension}
\alias{suggestExtension}
\title{Tries to locally add one link each and returns a list of links ordered by their chi-squared differences to the original model}
\usage{
suggestExtension(original_model, parallel = F, mc = 1,
  sample_range = c(10^(2:-1), 0, -10^(-1:2)),
  padjust_method = "bonferroni", print = F,
  fname = "Additional_link_suggestion.txt")
}
\arguments{
\item{original_model}{MRAmodel or MRAmodelSet object describing the model and its best fit, containing the data}

\item{parallel}{Boolean number indicating whether addition is executed in a parallel fashion}

\item{mc}{Number of cores that should be used for the computation}

\item{sample_range}{Numeric vector containing all starting values for the new link (DEFAULT: c(10^(2:-1),0,-10^(-1:2)))}

\item{padjust_method}{The method to use for the adjusted p-value, as defined in p.adjust. 'BY' by default which provides the FDR under general dependence assumption (conservative)}

\item{print}{Boolean indicating whether the result should be printed in a text file "Additional_link_suggestion.txt"}

\item{fname}{Name of the file where the results should be printed}
}
\description{
A new link found to be suitable by the modeller can then added by re-running the createModel function with the altered adjacency information.
Note that values assigned to be exactly 1 inidicate almost always an non-identifiable link, whose combined value is assigned to another node in the combination!
}
\examples{
\dontrun{
ext_list = suggestExtension(mramodel)
}
}
\seealso{
selectMinimalModel, createModel
}
\author{
Bertram Klinger \email{bertram.klinger@charite.de}
}
